requirements.txt
streamlit
speechrecognition
pyaudio
transformers
torch
pillow
python-dotenv
langchain
langgraph
huggingface_hub
langchain-community

